+++ console/src/metadata/sourcesUtils.ts
@@ -71,6 +71,23 @@ export const addSource = (
   const adaptedCustomizations = adaptCustomizations(payload?.customization);
 
   if (driver === 'mssql') {
+    if (payload?.replace_configuration) {
+      return {
+        type: 'mssql_update_source',
+        args: {
+          name: payload.name,
+          configuration: {
+            connection_info: {
+              connection_string: payload.dbUrl,
+              pool_settings: payload.connection_pool_settings,
+            },
+            read_replicas: replicas?.length ? replicas : null,
+          },
+          replace_configuration,
+          ...adaptedCustomizations,
+        },
+      };
+    }
     return {
       type: 'mssql_add_source',
       args: {
@@ -93,6 +110,22 @@ export const addSource = (
       typeof payload.dbUrl === 'string'
         ? JSON.parse(payload.dbUrl)
         : payload.dbUrl;
+    if (payload?.replace_configuration) {
+      return {
+        type: 'bigquery_update_source',
+        args: {
+          name: payload.name,
+          configuration: {
+            service_account,
+            global_select_limit: payload.bigQuery.global_select_limit,
+            project_id: payload.bigQuery.projectId,
+            datasets: payload.bigQuery.datasets.split(',').map(d => d.trim()),
+          },
+          replace_configuration,
+          ...adaptedCustomizations,
+        },
+      };
+    }
     return {
       type: 'bigquery_add_source',
       args: {
@@ -102,6 +135,32 @@ export const addSource = (
           global_select_limit: payload.bigQuery.global_select_limit,
           project_id: payload.bigQuery.projectId,
           datasets: payload.bigQuery.datasets.split(',').map(d => d.trim()),
+        },
+        replace_configuration,
+        ...adaptedCustomizations,
+      },
+    };
+  }
+
+  if (
+    (driver === 'postgres' || driver === 'citus') &&
+    payload?.replace_configuration
+  ) {
+    return {
+      type: `${driver === 'postgres' ? 'pg' : 'citus'}_update_source`,
+      args: {
+        name: payload.name,
+        configuration: {
+          connection_info: {
+            database_url: payload.connection_parameters
+              ? { connection_parameters: payload.connection_parameters }
+              : payload.dbUrl,
+            use_prepared_statements: payload.preparedStatements,
+            isolation_level: payload.isolationLevel,
+            pool_settings: payload.connection_pool_settings,
+            ssl_configuration: payload.sslConfiguration,
+          },
+          read_replicas: replicas?.length ? replicas : null,
         },
         replace_configuration,
         ...adaptedCustomizations,